Fix the chart converter + fix the pause screen
How I fixed it:
Instead of sorting each section and calling it there to translate a json to a cbin, I initialized a local array that captures every single unsorted note in each section, unsorted, and then puts all of it into the mixing of sorting and putting it all in the last phase which is to write it all into the .cbin file. The reason I wanted to do this was because 1. I wanted to finally fix it in the future and 2.