Major Change! MetaNote rework + lane revamp
This was a nightmare to be honest but was an interesting perspective as to why you don't mess with shit. I wasted most of my day on this but at least it's better than school!

At least...
It's currently broken right now and I'm trying to fucking fix it. I get the top 1065 bottom 0 bug and it's been pissing me off to the point of not caring anymore.