Decouple the note logic and the note rendering
Now, really, it is fully decoupled now.

I've also done a secret optimization which you'll see once I get that working.

But for now, it's half stable. It was flickering and notes can spawn right before it hovers the receptor for the same reason. But, I had fixed the greedyMergeTemp beforehand.

I also am determined to fix the 16x merged notes not rendering properly.