Put new note rendering logic in render function
Since we've fully decoupled the note logic and note rendering, I decided to now put all of the rendering logic inside the render function on Main.hx

Now a strong warning, do this. It'll not break botTimers. But, a new thing will emerge, and that is basically when the receptor is confirmed early before the notes. But, that doesn't matter anyway. So now, notes fully render in real time & can handle conditions where the receptor should've resetted but it's been delayed by one frame. Yes, the one frame delay was removed there for some reason when having the shit onto the rendering function now. Pretty interesting but fascinating adventure doing this one note system feature in the first place. I am so fucking happy right now that it works perfectly I can't believe it :feelsgood: 🤩