Eventually remove the now-unused properties in the `Note` class
They basically waste RAM anyway and they were used before in older `NoteSystem` code that I eventually had the innovation to quickly optimize & revamp so those are gone. And no, the Sustain class doesn't have to do anything now considering it can stay the way it was before. And, as a last thing, the amount of virtual note and sustain elements are typically very small which is good for fully decoupling note logic & rendering which in case is beneficial considering it's a huge win. Just fucking wow.