Fix note missing not muting the vocals
This also implements greedy note merging (up to x128 and x16 at minimum).

And yes this does infact take a toll on the cpu instead of the gpu (like switching things around and shit), so I'm going to be profiling it to find out what the fuck is going on.