Just only apply the restrict shit onto both chart_file externs
Just found out setting an element in memory-mapped io causes the performance to tank (setNote being the main cause because it's using it), so I came up with a pretty snazzy idea to just remove the 3 bits from MetaNote entirely and just have the 3 flags be stored via a bitset. Idk however.