Hell yeah, reduce map get calls for noteTypeFunctionality
Yeah do that to even optimize note processing. Yes, seriously, I just found out about map's getting being slow and now I did this lil fix to it.

Yeah I used an array instead of a map for `noteTypeFunctionality` because there are only so many values you can basically use on type anyway and found it to be very mischievous to make it a map in the end, so I was like "fuck it, let's do what chatgpt said" and reduce frequent use of this small map's getting for every note.

Turns out a Map is just very slow at getting. I wouldn't even get the element in such small map, if at all, if it's by a thing that has so little values in the first place, so I just basically turned it to a whole lil array.