Fix a bug on `FreeplayScreen` + chartconverter fix
Just one moment, I used claude.ai to do this. I got it to actually fix it, for real. This happens when I have 4 songs are shown in my `chapter1` data json instead of 3 songs which I was using until my curiousity got the best of me and I finally started to test the unpredictable chart conversion.

Though, this was committed after I got the unpredictable chart conversion fully working.