Alr new method time
It's okay because you're already going to be rendering so little elements anyway so figured I'd scale it to just this.

Thing is, it works perfectly. No artifacts or anything since the notes are already going to be buried by the greedy-merged notes like how you'll naturally bury regular notes with a chunk of notes anyway.